fix more post-refactoring problems
