+++ tests/org.eclipse.n4js.n4mf.xpect.ui.tests/testProjects/test.06/manifest.n4mf.xt
@@ -35,5 +35,5 @@ Sources {
 	}
 }
 
-// XPECT warnings --> "Extended runtime environment cannot be specified for test projects." at "ExtendedRuntimeEnvironment: test.data.rl.project.one"
+// XPECT warnings --> "Extended runtime environment cannot be specified for test projects." at "ExtendedRuntimeEnvironment"
 ExtendedRuntimeEnvironment: test.data.rl.project.one
