+++ tests/org.eclipse.n4js.n4mf.xpect.ui.tests/testProjects/test.10/manifest.n4mf.xt
@@ -38,8 +38,8 @@ Sources {
 	}
 }
 
-// XPECT warnings -->"Tested projects should have the same project type." at "TestedProjects { test.data.re.project, test.data.rl.project.one }"
 TestedProjects {
+	// XPECT warnings -->"Tested projects should have the same project type." at "test.data.re.project"
 	test.data.re.project,
 	test.data.rl.project.one
 }
